permission verif + cmd vote fixer et amelioré

+++ commands/moderation/say.js
@@ -5,6 +5,11 @@ module.exports = {
     description: "Faire parler Netbeans",
     usage: "<message>",
     run: async(client, message, args) => {
+        // Member doesn't have permissions
+        if (!message.member.hasPermission("MANAGE_MESSAGES")) {
+            return message.reply("Tu ne peux pas utiliser cette commande...").then(m => m.delete(5000));
+        }
+
         if (message.deletable) message.delete();
 
         if (args.length < 1)
